fix: abort streams with never ending reads
When calling `.return` on a generator object, it means the next call
to `.next` will result in `{ done: true }`.  If there's an in-flight
invocation of `.next`, that needs to resolve before the `.return`
will be processed.  If `.next` never resolves, `.return` will never
be processed, the `AbortError` will never be thrown and everything
hangs indefinitely.

The fix here is to not await on the result of `.return` and move on
to throwing the `AbortError`.  If `.return` rejects we should do
something with the error, so a new `onReturnError` callback is added
to the options to receive the error if the caller is interested.